requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
async-timeout==5.0.1
attrs==24.2.0
bidict==0.23.1
contourpy==1.3.1
cycler==0.12.1
dnspython==2.7.0
fonttools==4.55.3
frozenlist==1.5.0
h11==0.14.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.0
mpl-finance==0.10.1
multidict==6.1.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pandas-ta==0.3.14b0
pillow==11.0.0
propcache==0.2.1
pymongo==4.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
simple-websocket==1.1.0
six==1.17.0
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
wsproto==1.2.0
yarl==1.18.3
